Add files via upload
*Number Convertor
Write a function that converts a decimal number to binary and hexadecimal.

 
The given number will always be in decimal format. The "From" select menu will only have a Decimal option,
but the "To" select menu will have two options: Binary and Hexadecimal. 
This means that our program should have the functionality to convert decimal to binary and decimal to hexadecimal. When you convert to hexadecimal it must be upper case.
Note that the "To" select menu by default is empty. You have to insert the two options ('Binary' and 'Hexadecimal') inside before continuing. Also, they should have values ('binary' and 'hexadecimal').
•	When the [Convert it] button is clicked, the expected result should appear in the [Result] input field.