Add files via upload
List of Items
Write a function that reads the text inside an input field and appends the specified text to a list inside an HTML page.
Add / Delete
Extend the previous problem to display a [Delete] link after each list item. Clicking it should delete the item with no confirmation. You have to add href="#" to the link element.