Add files via upload
Time Converter
Create a program that converts different time units. Your task is to add a click event listener to all [CONVERT] buttons. When a button is clicked, read the corresponding input field, convert the value to the three other time units and display it in the input fields.
Example
 
 
One day is equal to 24 hours/1440 minutes/86400 seconds. Whichever button we click, the input fields should change depending on the added value on the left. (For example, if we write 48 hours and click convert the days, the field value should change to 2).
3.	Locked Profile
In this problem, you should create a JS functionality that shows and hides the additional information about users.
 
When one of the [Show more] buttons is clicked, the hidden information inside the div should
be shown, only if the profile is not locked! If the current profile is locked, nothing should happen.
 
If the hidden information is displayed and we lock the profile again, the [Hide it] button should not be working! 
Otherwise, when the profile is unlocked and we click on the [Hide it] button, the new fields must hide again.